move xgrammar related utils to backend_xgrammar.py
